==> 2_omnicomm_api/1_get_json.py <==
import requests
import json
from pprint import pprint


def main():
    def get_data(url, token):
        url = url
        payload={}
        headers = {
        'Content-Type': 'application/json',
        'Authorization': token
        }

        response = requests.request("GET", url, headers=headers, data=payload)
        return response.json()
        
        
        
   
    token = 'JWT eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6OTc2LCJsb2dpbiI6InBha2Vyc2VydmljZSIsInV1aWQiOiIyYWE1ZjgwZC05NWQ1LTM4YTAtOWVmZC1jMDMzNGMwOGNhNzYiLCJhdXRvY2hlY2tfaWQiOjgwOCwicGVybWlzc2lvbnMiOlsiI1VzZXIiLCIjdXNlcl9hZGQiLCJhc2UucmVwb3J0cy5tYXAyIiwiYXNlLnJlcG9ydHMubG9jYXRpb24iLCJhc2UucmVwb3J0cy5yZWZyaWdlcmF0b3J3b3JrIiwiYXNlLnJlcG9ydHMudGlyZXByZXNzdXJlIiwiYXNlLnJlZnJpZ2VyYXRvciIsImFzZS50cG1zIiwiYXNlLm1vZGJ1c2dlbiIsImFzZS5tb2RidXNsbHMiLCJhc2UuZ3JvdXBzLmRyaXZlci52aWV3IiwiYXNlLmdyb3Vwcy5yb3V0ZS52aWV3IiwiYXNlLnVzZXJyZXBvcnRzY29udHJvbCIsImFzZS5yZXBvcnRzLnBlcmlvZGljc2VydmljZSIsIm1hcHMueWFuZGV4Lm1hcCIsIm1hcHMud2lraS5tYXAiLCJtYXBzLm9zbS5tYXAiLCJtYXBzLm9tbmljb21tLm1hcCIsIm1hcHMub21uaWNvbW0uamFtcyIsImFzZS5yZXBvcnRzLmZ1ZWxsZXZlbHMiLCJhc2UuZ3JvdXBzLmdlb3pvbmUuY3VzdG9tIiwiYXNlLmdyb3Vwcy5nZW96b25lLnZpZXciLCJhc2UuZ3JvdXBzLmdlb3pvbmUuZWRpdHByb2ZpbGUiLCJzZXJ2aWNlLmZ1ZWxiYWxhbmNlIiwic2VydmljZS5yZXBvcnRzLmZ1ZWxzaGVldCIsImFzZS5yZXBvcnRzLmdyb3Vwc3RhdCIsImFzZS5yZXBvcnRzLnVuaXZlcnNhbCIsImlkZW50aWZpY2F0aW9uIiwibWFpbnBhZ2UubmV3IiwibWFwcy5nb29nbGUubWFwIiwibWFwcy5nb29nbGUuc2F0ZWxpdGUiLCJtYXBzLmdvb2dsZS5oeWJyaWQiLCJhc2UucmVwb3J0cy5ldmVudHMiLCJhc2UuZ3JvdXBzLnZlaGljbGUuY3VzdG9tIiwiYXNlLmdyb3Vwcy52ZWhpY2xlLnZpZXciLCJhc2UuZ3JvdXBzLnZlaGljbGUuZWRpdHRyZWUiLCJhc2UuZ3JvdXBzLnZlaGljbGUudmlld3Byb2ZpbGUiLCJhc2UuZ3JvdXBzLnZlaGljbGUuZWRpdGluZm8iLCJhc2UuZ3JvdXBzLnZlaGljbGUuZWRpdHByb2ZpbGUiLCJhc2UuZ3JvdXBzLnZlaGljbGUuc2VydmljZW1uZ21udCIsImFzZS5ncm91cHMudmVoaWNsZS5yZWNhbGMiLCJhc2UucmVwb3J0cy5nZW96b25lc3JlcG9ydCIsImFzZS5ncm91cHMuZ2Vvem9uZS5kZWxldGUiLCJhdXRoLmNsaWNrcmVwb3J0cyIsImF1dG9jaGVjay5hY2Nlc3MiLCJzZXJ2aWNlLmJpbGxpbmcubmV3IiwibWFwcy53aWtpLm1hcCIsIm1hcHMub3NtLm1hcCIsIm1hcHMuc3B1dG5pay5tYXAiLCJzZXJ2aWNlLmNvcHMucmVhZCIsInNlcnZpY2UuY29wcy51cGRhdGUiLCJzZXJ2aWNlLmNvcHMuYWRkIiwic2VydmljZS5DQU5fYnlfZW1kZCIsInNlcnZpY2Uub3ZtcyIsInNlcnZpY2UuaGZtcy5tZXNzYWdlcyIsInNlcnZpY2UucmVwb3J0cy5oZm1zX3N0YXR1cyIsInNlcnZpY2UucmVwb3J0cy5oZm1zX21lc3NhZ2VzIiwic2VydmljZS5jb3BzLnJlYWQiLCJzZXJ2aWNlLmNvcHMudXBkYXRlIiwic2VydmljZS5jb3BzLmFkZCIsInNlcnZpY2UuYmlsbGluZy51c2VyLmVkaXRvciIsInNlcnZpY2Uuc2FmZWRyaXZpbmdyZXBvcnQiLCJzZXJ2aWNlLmJpbGxpbmcubmV3IiwiYXNlLmZ1ZWxtYXNzIiwiZnVlbG1hc3MiLCJmdWVsLmJhbGFuY2UubmV3IiwiZnVlbC5iYWxhbmNlLmZlc19hcGkiLCJsbHM1LmFjY2VzcyIsImF1dGguY2xpY2tyZXBvcnRzIiwic2VydmljZS5DQU5fYnlfZW1kZCJdLCJkcml2ZXJncm91cF9pZCI6IjdkMWQ5ZGRiLTQxNmQtMzljZC04ZjhkLWFmNjQzZTU0MWUxOCIsInZlaGljbGVncm91cF9pZCI6ImVjMjIyZGVmLTg1N2YtM2M5Yy1hNjdhLTkwM2RkYjkxMGEyYSIsInVzZXJncm91cF9pZCI6bnVsbCwiZ2Vvem9uZWdyb3VwX2lkIjoiNWEyZGYwNGItODJkOS0zMWMwLTg0ZGUtNDdhMGRmZTQ3MDcxIiwicm91dGVncm91cF9pZCI6Ijg0NTBlNGM5LTlkNTctMzQyNS05NWFkLWQ0M2FjMTYzYjQ5YiIsInNlcnZlcl9uYW1lIjoiZ3JlZW5zY2FsZSIsIm1heF9yZXBvcnRfcGVyaW9kIjo2MCwic2VydmVyIjp7ImlkIjo2LCJob3N0IjoiaHR0cDovLzEwLjU1LjYuMTE2IiwicG9ydCI6ODA4MCwic2x1ZyI6ImdyZWVuc2NhbGUiLCJmcWRuIjoiMTAuNTUuNi4xMTYiLCJ2b2x1bWVfdW5pdHMiOiJMIiwibGlnaHQiOjB9LCJyb2xlcyI6W3siaWQiOjQsIm5hbWUiOiJVc2VyIiwic2x1ZyI6InVzZXIiLCJ3ZWlnaHQiOjEwfSx7ImlkIjo2OCwibmFtZSI6InVzZXJfYWRkIiwic2x1ZyI6InVzZXJfYWRkIiwid2VpZ2h0IjowfV0sIndsIjp7fSwiaWF0IjoxNjY5ODc3NjQ2LCJleHAiOjE2Njk4ODEyNDYsImF1ZCI6ImFzZSJ9.-zKLizKapj0qcEfS_fb5r8Ah--QOu8e4kFdXEHXMbCQ'
    units = get_data('https://online.omnicomm.ru/ls/api/v1/tree/vehicle', token)
    online_data = get_data('https://online.omnicomm.ru/ls/api/v2/activity/vehicles', token)
    
    json.dump(units, open( "JSON_om_units.json", 'w' ))
    json.dump(online_data, open( "JSON_om_online.json", 'w' ))
if __name__ == '__main__':
    main()
